pyproject.toml
[tool.poetry]
name = "manual_influencer"
version = "0.1.0"
description = ""
authors = ["Rodrigo Bermudez Schettino <rodrigobdz@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
google-cloud-texttospeech = "^2.25.1"
google-genai = "^1.8.0"
google-cloud-storage = "^3.1.0"
opencv-python = "^4.11.0.86"
imageio-ffmpeg = "^0.6.0"
ffmpeg = "^1.4"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
